requirements.txt
boto3==1.34.34
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
schedule==1.2.0
streamlit==1.30.0
pandas==2.1.4
plotly==5.18.0

